+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Exchange backup issue caused by incorrect json serialization
 
 ### Changed
-- Do not display all the items that we restored at the end if there are more than 15
+- Do not display all the items that we restored at the end if there are more than 15. You can override this with `--verbose`.
 
 ### Known Issues